30(4)

+++ srcs/type_2.c
@@ -52,8 +52,8 @@ void	f_type_u(char **res, unsigned long long input)
 {
 	char *str;
 
-	// printf("f_type_u-\t[%llu]\n", input);
-	str = ft_itoa_u((unsigned long long)input);
+	// printf("f_type_u-\t[%hu]\n", input);
+	str = ft_itoa_u(input);
 	f_join_free(res, str);
 	free(str);
 }
